// src/shared/generated/MacroCreateInputRecord.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MacroRepeat } from "./MacroRepeat";
import type { MacroShortcutSourceScope } from "./MacroShortcutSourceScope";
import type { MacroStepInputRecord } from "./MacroStepInputRecord";
import type { MacroTrigger } from "./MacroTrigger";

export type MacroCreateInputRecord = { enabled?: boolean, activationMode?: string, name: string, roleIds: Array<string>, shortcutSourceScope?: MacroShortcutSourceScope, trigger?: MacroTrigger, repeat?: MacroRepeat, steps: Array<MacroStepInputRecord>, };